Small refactoring and enhancements
